Simplify SimpsonsRule integration calucation.
Simpson's Rule now uses a simpler method to integrate rather
than the overlycomplex one in the previous commit.

This also fixes a bug in the previous commit (the actual formula was wrong).

Hopefully, loss of precision is reduced in this commit, but likely not.

Adds an extra test testing Simpson's Rule at low interval counts.